requirements.txt
# TradingView Webhook自动交易系统依赖包
# 适用于Python 3.11+

# Web框架
Flask==2.3.3
Werkzeug==2.3.7

# OKX官方API SDK
okx==2.1.2

# 环境变量管理
python-dotenv==1.0.0

# 日志处理 (Python内置，无需安装)
# logging==0.4.9.6

# 网络请求
requests==2.31.0
urllib3==2.0.7

# JSON处理 (Python内置，无需安装)
# jsonschema==4.19.0

# 时间处理
pytz==2023.3

# 数据处理
pandas==2.1.1
numpy==1.25.2

# 密码学（用于签名验证）
cryptography==41.0.4

# 可选：通知功能
# 微信机器人
# wechatpy==1.8.18

# 可选：邮件通知
# smtplib（Python内置）

# 开发工具（可选）
pytest==7.4.2
pytest-flask==1.2.0
